pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"


[project]
name = "dota_banpick"
authors = [{name = "Sukai Huang", email = "huangsukai1997@gmail.com"}]
version = "0.1.0"
requires-python = ">=3.10"
description = "BP helper for DOTA 2"
readme = "readme.md"
dependencies = [
    "requests","pandas","numpy", "undetected-chromedriver",
    "tqdm","ipywidgets", "hydra-core",
    "jupyterlab","selenium","django", "natsort", "Pillow", "streamlit"
]

[project.optional-dependencies]
dev = [
    "pytest"
]

[tool.setuptools.packages.find]
exclude = ["config", "outputs", "multirun", "logs", "frontend", "Value_resources"] 

[project.scripts]
dota-banpick = "dota_banpick:main_cli"